// maxxinke-admin/src/types/system.ts
/**
 * 联系方式信息接口
 */
export interface ContactInfo {
  tel: string;        // 电话
  mobile: string;     // 手机
  email: string;      // 邮箱
  serviceEmail: string; // 客服邮箱
  address: string;    // 地址
  postcode: string;   // 邮编
  website: string;    // 网站
  wechat: string;     // 微信公众号
} 